requirements.txt
Django==4.1
djangorestframework==3.13.1
django-extensions==3.2.0
django-filter==22.1
ipdb==0.13.9
